+++ src/main/java/minesweeper/banner/StatusIndicator.java
