+++ src/main/java/minesweeper/banner/StatusIndicatorClickedListener.java
