+++ src/main/resources/bomb.png
