+++ src/main/resources/lose.png
