+++ src/main/resources/win.png
